refactor: extract service logic to base service (API-77) (#24)
This refactors the current model services and extracts the shared logic to a base service. The new base service can be used across any new services without needing to duplicate content.
